Renamed some files for cosistancy
